+++ resources/js/Components/VideoPlayer/OttTopRightDisplay/NowPlayingInfo.vue
@@ -58,9 +58,6 @@ import { useVideoPlayerStore } from "@/Stores/VideoPlayerStore.js"
 import { useStreamStore } from "@/Stores/StreamStore"
 import { useChatStore } from "@/Stores/ChatStore.js"
 import { useUserStore } from "@/Stores/UserStore"
-import VideoOTTChat from "@/Components/VideoPlayer/Chat/VideoOTTChat"
-import Channels from "@/Components/VideoPlayer/Channels/Channels"
-import ChannelFooter from "@/Components/VideoPlayer/Channels/ChannelFooter.vue";
 import SingleImage from "@/Components/Multimedia/SingleImage.vue";
 
 let videoPlayerStore = useVideoPlayerStore()