+++ resources/js/Stores/WelcomeStore.js
@@ -1,7 +1,7 @@
 import { defineStore } from "pinia";
 import {ref} from "vue";
 
-export let useWelcomeStore = defineStore('welcomeStore', {
+export const useWelcomeStore = defineStore('welcomeStore', {
     state() {
         return {
             showLogin: Boolean,